rename CdoFlavoredParser to RedactableMarkdownParser
